# src/main/java/com/genesis/proyecto2/dtos/UpdateExchangeRateRequest.java
package com.genesis.proyecto2.dtos;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class UpdateExchangeRateRequest {
    private Double copPerUsd;
}