# src/main/java/com/genesis/proyecto2/dtos/UserAdminView.java
package com.genesis.proyecto2.dtos;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class UserAdminView {
    private Long id;
    private String email;
    private Integer balance;
    private String activePlan;
    private String status;
}