# src/main/java/com/genesis/proyecto2/dtos/UserProfileResponse.java
package com.genesis.proyecto2.dtos;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class UserProfileResponse {
    private Long id;
    private String name;
    private String email;
    private Integer balance;
    private PlanInfo activePlan;
    private String status;
}